Finance ECF uses persisted statements
