+++ app/controllers/finance/ecf/contracts_controller.rb
@@ -5,6 +5,11 @@ module Finance
     class ContractsController < BaseController
       def show
         @ecf_lead_provider = lead_provider_scope.find(params[:id])
+        # TODO: back link goes to latest statement rather that where they were
+        @latest_statement = Statement::ECF
+          .where(cpd_lead_provider: @ecf_lead_provider.cpd_lead_provider)
+          .order(deadline_date: :desc)
+          .first
         @contract = @ecf_lead_provider.call_off_contract
       end
