+++ app/views/finance/ecf/contracts/show.html.erb
@@ -1,3 +1,3 @@
-<% content_for :before_content, govuk_back_link(text: "Back", href: finance_ecf_payment_breakdown_path(id: @ecf_lead_provider.id)) %>
+<% content_for :before_content, govuk_back_link(text: "Back", href: finance_ecf_payment_breakdown_statement_path(payment_breakdown_id: @ecf_lead_provider.id, id: @latest_statement.name)) %>
 
 <%= render Finance::ECF::Contract.new(contract: @contract) %>
